+++ src/com/troblecodings/guilib/ecs/entitys/render/UILines.java
@@ -7,15 +7,15 @@ import com.troblecodings.guilib.ecs.entitys.UIComponent;
 
 public class UILines extends UIComponent {
 
-    private final double[] lines;
-    private final double[] lineCache;
+    private final float[] lines;
+    private final float[] lineCache;
     private final float width;
     private int color;
 
-    public UILines(final double[] lines, final float width) {
+    public UILines(final float[] lines, final float width) {
         super();
         this.lines = Arrays.copyOf(lines, lines.length);
-        this.lineCache = new double[lines.length];
+        this.lineCache = new float[lines.length];
         this.width = width;
         this.color = 0xFFFFFFFF;
     }
@@ -29,8 +29,8 @@ public class UILines extends UIComponent {
     @Override
     public void update() {
         for (int i = 0; i < lines.length; i += 2) {
-            this.lineCache[i] = this.lines[i] * parent.getWidth();
-            this.lineCache[i + 1] = this.lines[i + 1] * parent.getHeight();
+            this.lineCache[i] = this.lines[i] * (float) parent.getWidth();
+            this.lineCache[i + 1] = this.lines[i + 1] * (float) parent.getHeight();
         }
     }
